chain: do mtp and too-new checks in spv as well.
Rename chain-test to chain-full-test.
Rename chain-reorg-test to chain-test.
