+++ lib/blockchain/chain.js
@@ -475,11 +475,6 @@ class Chain extends AsyncEmitter {
         100);
     }
 
-    // Skip all blocks in spv mode once
-    // we've verified the network target.
-    if (this.options.spv)
-      return this.state;
-
     // Ensure the timestamp is correct.
     const mtp = await this.getMedianTime(prev);
 
@@ -500,6 +495,11 @@ class Chain extends AsyncEmitter {
         0,
         true);
     }
+
+    // Skip all blocks in spv mode once
+    // we've verified the network target.
+    if (this.options.spv)
+      return this.state;
 
     // Calculate height of current block.
     const height = prev.height + 1;
